+++ packages/core/src/di/injection_token.ts
@@ -60,13 +60,17 @@ export class InjectionToken<T> {
     providedIn?: Type<any>| 'root' | null,
     factory: () => T
   }) {
-    if (options !== undefined) {
+    this.ngInjectableDef = undefined;
+    if (typeof options == 'number') {
+      // This is a special hack to assign __NG_ELEMENT_ID__ to this instance.
+      // __NG_ELEMENT_ID__ is Used by Ivy to determine bloom filter id.
+      // We are using it to assign `-1` which is used to identify `Injector`.
+      (this as any).__NG_ELEMENT_ID__ = options;
+    } else if (options !== undefined) {
       this.ngInjectableDef = defineInjectable({
         providedIn: options.providedIn || 'root',
         factory: options.factory,
       });
-    } else {
-      this.ngInjectableDef = undefined;
     }
   }
