+++ packages/core/src/di/scope.ts
@@ -6,7 +6,6 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {Type} from '../interface/type';
 import {InjectionToken} from './injection_token';
