+++ packages/core/src/render/api.ts
@@ -13,6 +13,7 @@ import {injectRenderer2 as render3InjectRenderer2} from '../render3/view_engine_
 import {noop} from '../util/noop';
 
 
+
 /**
  * @deprecated Use `RendererType2` (and `Renderer2`) instead.
  * @publicApi
