+++ packages/core/src/render3/instructions.ts
@@ -6,10 +6,8 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
+import {InjectFlags, InjectionToken, Injector} from '../di';
 import {resolveForwardRef} from '../di/forward_ref';
-import {InjectionToken} from '../di/injection_token';
-import {Injector} from '../di/injector';
-import {InjectFlags} from '../di/injector_compatibility';
 import {Type} from '../interface/type';
 import {QueryList} from '../linker';
 import {Sanitizer} from '../sanitization/security';
