+++ packages/core/src/render3/interfaces/injector.ts
@@ -7,8 +7,9 @@
  */
 
 import {InjectionToken} from '../../di/injection_token';
-import {InjectFlags} from '../../di/injector_compatibility';
+import {InjectFlags} from '../../di/interface';
 import {Type} from '../../interface/type';
+
 import {TElementNode} from './node';
 import {LView, TData} from './view';
